refactor: removes duplicate files, refactor UserService
